pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.54.0",
    "chardet>=5.2.0",
    "google-genai>=1.21.1",
    "openai>=1.90.0",
    "pillow>=11.2.1",
    "pypdf2>=3.0.1",
    "python-magic>=0.4.27",
    "streamlit>=1.46.0",
]
